+++ fukkatsu/utils/medic.py
@@ -3,8 +3,9 @@ import os
 import openai
 
 from fukkatsu.utils.prompt import (ADDITIONAL, CONTEXT, CONTEXT_MUTATE,
-                                   OUTPUT_CONSTRAINTS,
-                                   OUTPUT_CONSTRAINTS_MUTATE)
+                                   CONTEXT_TWIN, OUTPUT_CONSTRAINTS,
+                                   OUTPUT_CONSTRAINTS_MUTATE,
+                                   OUTPUT_CONSTRAINTS_TWIN)
 
 try:
     openai.api_key = os.environ.get("OPENAI_API_KEY")
@@ -13,11 +14,14 @@ except:
     print("OPENAI_API_KEY not found in environment variables.")
     raise Exception("OPENAI_API_KEY not found in environment variables.")
 
-MODEL = "gpt-3.5-turbo"
-
 
 def defibrillate(
-    inputs: str, faulty_function: str, error_trace: str, additional_req: str = ""
+    inputs: str,
+    faulty_function: str,
+    error_trace: str,
+    model: str,
+    temperature: float,
+    additional_req: str = "",
 ) -> str:
     if additional_req == "":
         set_prompt = (
@@ -32,14 +36,14 @@ def defibrillate(
         )
 
     response = openai.ChatCompletion.create(
-        model=MODEL,
+        model=model,
         messages=[
             {"role": "system", "content": set_prompt},
         ],
         max_tokens=1024,
         n=1,
         stop=None,
-        temperature=0.1,
+        temperature=temperature,
     )
 
     corrected_function = response["choices"][0]["message"]["content"].strip()
@@ -47,21 +51,50 @@ def defibrillate(
     return corrected_function
 
 
-def enhance(inputs: str, target_function: str, request: str = "") -> str:
+def enhance(
+    inputs: str, target_function: str, model: str, temperature: float, request: str = ""
+) -> str:
     set_prompt = (
         f"{CONTEXT_MUTATE}\n\n{target_function}\n\nThe function received the following inputs:\n\n"
         f"{inputs}\n\nThe user requests the following:\n{request}\n{OUTPUT_CONSTRAINTS_MUTATE}"
     )
 
     response = openai.ChatCompletion.create(
-        model=MODEL,
+        model=model,
         messages=[
             {"role": "system", "content": set_prompt},
         ],
         max_tokens=1024,
         n=1,
         stop=None,
-        temperature=0.1,
+        temperature=temperature,
+    )
+
+    mutated_function = response["choices"][0]["message"]["content"].strip()
+
+    return mutated_function
+
+
+def twin(
+    inputs: str,
+    target_function: str,
+    model: str,
+    temperature: float,
+) -> str:
+    set_prompt = (
+        f"{CONTEXT_TWIN}\n\n{target_function}\n\nThe function received the following inputs:\n\n"
+        f"{inputs}\n\n{OUTPUT_CONSTRAINTS_TWIN}"
+    )
+
+    response = openai.ChatCompletion.create(
+        model=model,
+        messages=[
+            {"role": "system", "content": set_prompt},
+        ],
+        max_tokens=1024,
+        n=1,
+        stop=None,
+        temperature=temperature,
     )
 
     mutated_function = response["choices"][0]["message"]["content"].strip()